First stab at retrieving Drone config data.
The ARDrone responds to a AT*CTRL=4 command by connecting back to
the client on port 5559 and sending a parameter list of settings
in the form of:

general:num_version_config       = 1
general:num_version_mb           = 33
general:num_version_soft         = 2.1.18
[etc]

as documented in section 8.1.2 (page 63 & 64) of the ARDrone
Developer Guide SDK 2.0.

This initial code starts a thread to accept connections and then
just dumps out the data passed back as is, for examination.